refactor(claimer): move broker mock to listener
